services/complaints-mcp/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Mail==0.9.1
pydantic==2.5.2
Werkzeug==2.3.7
pyOpenSSL==23.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
email-validator==2.1.0